pyproject.toml
[tool.poetry]
authors = ["XieJiSS <panrz@deepseek.com>"]
description = "Runch: refined munch. Attribute-style access for python dicts, with generic typing support. Help handle your configs, data, and more."
license = "Apache-2.0"
name = "runch"
packages = [{include = "runch"}]
readme = "README.md"
repository = "https://github.com/XieJiSS/runch.git"
version = "1.3.0"

[tool.poetry.dependencies]
asyncer = "^0.0.8"
pydantic = "^2.9.2"
python = "^3.12"
datamodel-code-generator = "^0.26.5"
runch-mergedeep = "^1.3.5"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pyright = "^1.1.388"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
